requirements.txt
aiogram>=3.14
telethon
sqlalchemy>=2.0
aiosqlite
google-generativeai
python-dotenv
pydantic>=2.0
pydantic-settings
apscheduler
